feat: Bean Factory 추상화
