+++ src/main/java/core/annotation/ComponentScan.java
@@ -1,5 +1,7 @@
 package core.annotation;
 
+import org.springframework.core.annotation.AliasFor;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -8,7 +10,9 @@ import java.lang.annotation.Target;
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 public @interface ComponentScan {
+    @AliasFor("basePackages")
     String[] value() default {};
 
+    @AliasFor("value")
     String[] basePackages() default {};
 }
