+++ src/main/java/core/mvc/tobe/BeanScanner.java
@@ -1,17 +1,17 @@
 package core.mvc.tobe;
 
-import com.google.common.collect.Sets;
 import core.annotation.Repository;
 import core.annotation.Service;
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
-import core.di.factory.BeanFactory;
+import core.di.factory.ComponentBeanFactory;
 import core.mvc.tobe.support.ArgumentResolver;
 import core.mvc.tobe.support.HttpRequestArgumentResolver;
 import core.mvc.tobe.support.HttpResponseArgumentResolver;
 import core.mvc.tobe.support.ModelArgumentResolver;
 import core.mvc.tobe.support.PathVariableArgumentResolver;
 import core.mvc.tobe.support.RequestParamArgumentResolver;
+import core.util.ReflectionUtils;
 import org.reflections.Reflections;
 import org.reflections.scanners.MethodAnnotationsScanner;
 import org.reflections.scanners.SubTypesScanner;
@@ -21,13 +21,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
 import org.springframework.core.ParameterNameDiscoverer;
 
-import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import static java.util.Arrays.asList;
 
@@ -43,30 +41,22 @@ public class BeanScanner {
                 new ModelArgumentResolver()
         );
 
-    private BeanFactory beanFactory;
+    private ComponentBeanFactory componentBeanFactory;
 
     public BeanScanner(Object... basePackage) {
         Reflections reflections = new Reflections(basePackage, new TypeAnnotationsScanner(), new SubTypesScanner(), new MethodAnnotationsScanner());
-        beanFactory = new BeanFactory(getTypesAnnotatedWith(reflections, Controller.class, Service.class, Repository.class));;
-        beanFactory.initialize();
+        componentBeanFactory = new ComponentBeanFactory();
+        componentBeanFactory.register(ReflectionUtils.getTypesAnnotatedWith(reflections, Controller.class, Service.class, Repository.class));
     }
 
     private static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
 
     public Map<HandlerKey, HandlerExecution> scanController() {
         Map<HandlerKey, HandlerExecution> handlers = new HashMap<>();
-        beanFactory.getControllers()
-                   .forEach(controller -> addHandlerExecution(handlers, controller, controller.getClass().getMethods()));
+        componentBeanFactory.getControllers()
+                            .forEach(controller -> addHandlerExecution(handlers, controller, controller.getClass().getMethods()));
 
         return handlers;
-    }
-
-    private Set<Class<?>> getTypesAnnotatedWith(Reflections reflections, Class<? extends Annotation>... annotations) {
-        Set<Class<?>> beans = Sets.newHashSet();
-        for (Class<? extends Annotation> annotation : annotations) {
-            beans.addAll(reflections.getTypesAnnotatedWith(annotation));
-        }
-        return beans;
     }
 
     private void addHandlerExecution(Map<HandlerKey, HandlerExecution> handlers, final Object target, Method[] methods) {
