+++ src/main/java/core/util/ReflectionUtils.java
@@ -1,11 +1,15 @@
 package core.util;
 
+import com.google.common.collect.Sets;
+import org.reflections.Reflections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.util.Set;
 
 public class ReflectionUtils {
 
@@ -81,6 +85,14 @@ public class ReflectionUtils {
             }
         }
         return false;
+    }
+
+    public static Set<Class<?>> getTypesAnnotatedWith(Reflections reflections, Class<? extends Annotation>... annotations) {
+        Set<Class<?>> beans = Sets.newHashSet();
+        for (Class<? extends Annotation> annotation : annotations) {
+            beans.addAll(reflections.getTypesAnnotatedWith(annotation));
+        }
+        return beans;
     }
 
 }
